agent-architecture/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiolimiter==1.2.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.79.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.1
fastapi==0.128.8
ffmpeg-python==0.2.0
filelock==3.19.1
frozenlist==1.8.0
fsspec==2025.10.0
future==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.83
langchain-text-splitters==0.3.11
langsmith==0.4.37
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.7.1
numpy==2.0.2
orjson==3.11.5
packaging==25.0
pillow==11.3.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.2
shellingham==1.5.4
sniffio==1.3.1
starlette==0.49.3
tenacity==9.1.2
tokenizers==0.22.2
tqdm==4.67.3
typer==0.23.2
typer-slim==0.23.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.39.0
voyageai==0.3.7
yarl==1.22.0
zstandard==0.25.0
